requirements.txt
streamlit>=1.56.0
google-genai>=1.73.0
python-dotenv>=1.0.0
